Refactor result view and add override for vocab
